+++ tests/testthat/test-MUS.R
@@ -26,7 +26,7 @@ test_that(desc = "(id: f12-v0.5.4-t1) Test Sample sizes for poisson distribution
   for (j in seq_along(m)) {
     for (i in seq_along(e)) {
       ss_MUS[i, j] <- MUS::MUS.calc.n.conservative(confidence.level = 0.95, tolerable.error = m[j], expected.error = e[i], book.value = sum(data$book.value)) # MUS::calc.n.conservative is used in MUS::MUS.planning()
-      ss_jfa[i, j] <- jfa::planning(conf.level = 0.95, materiality = m[j] / sum(data$book.value), expected = e[i] / sum(data$book.value), likelihood = "poisson")$n
+      ss_jfa[i, j] <- planning(conf.level = 0.95, materiality = m[j] / sum(data$book.value), expected = e[i] / sum(data$book.value), likelihood = "poisson")$n
     }
   }
 
@@ -57,7 +57,7 @@ test_that(desc = "(id: f12-v0.5.4-t2) Test most likely error and upper bound usi
       audited <- cbind(extract$sample, audit.value = audited)
       # Evaluate the sample, cache and print it
       e_MUS <- MUS::MUS.evaluation(extract = extract, filled.sample = audited)
-      e_jfa <- jfa::evaluation(conf.level = 0.95, materiality = m[j] / sum(data$book.value), method = "stringer", data = audited, values = "book.value", values.audit = "audit.value")
+      e_jfa <- evaluation(conf.level = 0.95, materiality = m[j] / sum(data$book.value), method = "stringer.binomial", data = audited, values = "book.value", values.audit = "audit.value")
       mle_MUS[i, j] <- as.numeric(e_MUS$MLE.final / sum(data$book.value))
       ub_MUS[i, j] <- as.numeric(e_MUS$UEL.low.error.rate / sum(data$book.value))
       mle_jfa[i, j] <- e_jfa$mle