+++ tests/testthat/test-other.R
@@ -18,6 +18,6 @@ context("Validation of other functionality")
 # jfa version 0.5.0
 
 test_that(desc = "(id: f4-v0.5.0-t1) Function test .markdown_call()", {
-  x <- jfa:::.markdown_call("rmarkdown::render")
+  x <- .markdown_call("rmarkdown::render")
   expect_equal(length(x), 1)
 })