+++ tests/testthat/test-samplingbook.R
@@ -26,7 +26,7 @@ test_that(desc = "(id: f15-v0.6.3-t1) Validate hypergeometric 99% upper bounds",
   for (N in seq(400, 1000, 50)) {
     for (n in seq(20, 300, 10)) {
       for (k in 0:10) {
-        ub_jfa[index] <- jfa::evaluation(materiality = 0.99, n = n, x = k, method = "hypergeometric", N.units = N, conf.level = ub_level, alternative = "less")$ub
+        ub_jfa[index] <- evaluation(materiality = 0.99, n = n, x = k, method = "hypergeometric", N.units = N, conf.level = ub_level, alternative = "less")$ub
         ub_sb[index] <- suppressWarnings(expr = {
           samplingbook::Sprop(m = k, n = n, N = N, level = interval_level)$ci$exact[2]
         })
@@ -46,7 +46,7 @@ test_that(desc = "(id: f15-v0.6.3-t2) Validate hypergeometric 95% upper bounds",
   for (N in seq(400, 1000, 50)) {
     for (n in seq(20, 300, 10)) {
       for (k in 0:10) {
-        ub_jfa[index] <- jfa::evaluation(materiality = 0.99, n = n, x = k, method = "hypergeometric", N.units = N, conf.level = ub_level, alternative = "less")$ub
+        ub_jfa[index] <- evaluation(materiality = 0.99, n = n, x = k, method = "hypergeometric", N.units = N, conf.level = ub_level, alternative = "less")$ub
         ub_sb[index] <- suppressWarnings(expr = {
           samplingbook::Sprop(m = k, n = n, N = N, level = interval_level)$ci$exact[2]
         })
@@ -66,7 +66,7 @@ test_that(desc = "(id: f15-v0.6.3-t3) Validate hypergeometric 90% upper bounds",
   for (N in seq(400, 1000, 50)) {
     for (n in seq(20, 300, 10)) {
       for (k in 0:10) {
-        ub_jfa[index] <- jfa::evaluation(materiality = 0.99, n = n, x = k, method = "hypergeometric", N.units = N, conf.level = ub_level, alternative = "less")$ub
+        ub_jfa[index] <- evaluation(materiality = 0.99, n = n, x = k, method = "hypergeometric", N.units = N, conf.level = ub_level, alternative = "less")$ub
         ub_sb[index] <- suppressWarnings(expr = {
           samplingbook::Sprop(m = k, n = n, N = N, level = interval_level)$ci$exact[2]
         })