+++ NEWS.md
@@ -6,6 +6,7 @@
 
 **Minor changes**
 
+- In the `evaluation()` function, the typical stringer bound has been changed from `method = "stringer"` to `method = "stringer.binomial"`. Furthermore, additional methods `stringer.poisson` and `stringer.hypergeometric` are added that apply the stringer bound using the Poisson and hypergeometric distributions, respectively. The `stringer` method will remain for now but redirect to `stringer.binomial`.
 - Compatibility with `rstantools 2.3.1`.
 
 # jfa 0.6.7